COS-5558: Create new "Add LinkedIn Connection" flow action node
