+++ packages/apps/frontera/src/routes/flow-editor/src/nodes/actions/index.ts
@@ -1,0 +1,2 @@
+export { EmailActionNode } from './EmailActionNode';
+export { SendConnectionRequestActionNode } from './SendConnectionRequestActionNode';
