+++ packages/apps/frontera/src/store/Flows/types.ts
@@ -1,6 +1,7 @@
 export enum FlowActionType {
   EMAIL_NEW = 'EMAIL_NEW',
   EMAIL_REPLY = 'EMAIL_REPLY',
+  SEND_CONNECTION_REQUEST = 'SEND_CONNECTION_REQUEST',
   FLOW_START = 'FLOW_START',
   FLOW_END = 'FLOW_END',
 }
